Intentos de arreglos

+++ Cursus/Libft/code/ft_memcpy.c
@@ -14,9 +14,9 @@
 
 void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
-	unsigned long	i;
-	char			*d;
-	char			*s;
+	size_t	i;
+	char	*d;
+	char	*s;
 
 	d = (char *) dst;
 	s = (char *) src;
